refactor(sender): SenderController 拆分为 sender 包
- controller/sender/controller.py - SenderController (~200行)
- controller/sender/workers.py - SendTaskWorker + QueueWorker (~200行)
- 删除旧的 sender_controller.py
- 使用绝对导入代替相对导入
