+++ src/danmaku_sender/ui/sender/page.py
@@ -15,7 +15,7 @@ from .components.dialogs.task_detail import TaskDetailDialog
 from danmaku_sender.ui.editor import EditorDialog
 from danmaku_sender.ui.monitor_dialog import MonitorDialog
 from danmaku_sender.ui.framework.style_loader import SvgIcon
-from danmaku_sender.controller.sender_controller import SenderController
+from danmaku_sender.controller.sender import SenderController
 from danmaku_sender.service.danmaku_parser import DanmakuParser
 from danmaku_sender.types.models.queue import QueueTask, TaskStatus
 from danmaku_sender.repo.history_manager import HistoryManager
